(SCHEMA) Update `returnKind` and associated properties
Prior to this change, the `returnKind` shared definition set a default
value. This was inaccurate, because `set.return` has different behavior
when `return` isn't defined - DSC invokes the **Get** operation to
construct the result after the **Set** operation for the resource
concludes.

This change removes the default from the shared definition, adds it
to the `test.return` property, and clarifies the behavior for the
`set.return` property. Finally, this change also makes `whatIf.return`
mandatory, as the synthetic after-state doesn't make sense for this
operation, even if DSC will technically accept it.